refactor: utils하위 파일명 lottoGameHandler를 common으로변경
