+++ src/view/InputWinningNumberView.js
@@ -1,6 +1,6 @@
 const { Console } = require('@woowacourse/mission-utils');
 const { isValidateNumber, isLottoRange, isDuplicate } = require('../utils/validation.js');
-const { separateStringBySpecificCharacter } = require('../utils/lottoGameHandler.js');
+const { separateStringBySpecificCharacter } = require('../utils/common.js');
 
 class InputWinningNumberView {
   isValidateWinningNumber(winningNumber) {
@@ -26,10 +26,10 @@ class InputWinningNumberView {
       throw new Error('[ERROR] 보너스 번호는 당첨 번호와 중복되지 않아야 합니다.');
   }
 
-  inputBonusNumberFromUser(judgePrize) {
+  inputBonusNumberFromUser(judgePurchasedLottoOfResult) {
     Console.readLine('보너스 번호를 입력해 주세요.\n', (bonusNumber) => {
       this.isValidateBonusNumber(bonusNumber);
-      judgePrize(this.winningNumber, bonusNumber);
+      judgePurchasedLottoOfResult(this.winningNumber, bonusNumber);
     });
   }
 }
